+++ CHANGELOG.md
@@ -31,6 +31,7 @@ IMPROVEMENTS:
    fields [GH-1567]
  * cli: `vault write -field` now allows selecting wrapped response fields
    [GH-1567]
+ * core: Response wrapping is now enabled for login endpoints [GH-1588]
  * credential/aws-ec2: Added a new constraint, 'bound_account_id' to the role
    [GH-1523]
  * physical/etcd: Support `ETCD_ADDR` env var for specifying addresses [GH-1576]